requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Vector Database & Search
chromadb==0.4.22
sentence-transformers==2.3.1

# Encryption
cryptography==42.0.0

# Desktop GUI
PySide6==6.6.1

# File Parsing
python-frontmatter==1.1.0
markdown==3.5.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
mypy==1.8.0
types-markdown==3.5.0.3
